% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_rOceans_2_global patterns
%   biodiversity v.0.0.6.r
\name{oceanDataCheck}
\alias{oceanDataCheck}
\title{ROxygen2 block for function #2.1
Function 2.1 oceanDataCheck: merges, checks and filters big datasets from OBIS and GBIF
merges, checks and filters big datasets from OBIS and GBIF}
\usage{
oceanDataCheck(source = "OBIS", GBIF_occurrences, OBIS_occurrences,
  remove_duplicates = T, remove_NAs = T, remove_land_occ = T)
}
\arguments{
\item{source}{source of occurrences, either "OBIS", "GBIF, or "GBIF_&_OBIS"}

\item{GBIF_occurrences}{a data.frame of downloaded occurrences from GBIF}

\item{OBIS_occurrences}{a data.frame of downloaded occurrences from ONIS}

\item{remove_duplicates}{indicates if checking and removing duplicates}

\item{remove_NAs}{indicates if checking and removing NAs}

\item{remove_NAs}{indicates if checking and removing land based occurrences  (non marine)}
}
\value{
Object of class data.frame with filtered occurrences
}
\description{
merges, checks and filters big datasets from OBIS and GBIF
}
\details{
The function merges, checks, and filters occurrences data downloaded from OBIS and GBIF
}
