% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Functions_rOcean_1_biodiversity_for_few_species_v0.1.0.r
\name{oceanDivDat}
\alias{oceanDivDat}
\title{Function 1.1 oceanDist: allows getting presences and distribution maps  for a list of species
ROxygen2 block for function #1
Allows getting distribution data  for a list of species from OBIS & GBIF}
\usage{
oceanDivDat(species_names, data_source = "OBIS", remove_duplicates = T,
  remove_land_dots = T, minimum_data = 5)
}
\arguments{
\item{species_names}{list of species names in latin}

\item{data_source}{source data, it can be "OBIS","GBIF" or both "OBIS_GBIF"}

\item{remove_duplicates}{if True, checks and removes duplicates}

\item{remove_land_dots}{if True, checks and removes land based occurrences}

\item{minimum_data}{treshold of minimum required data for including species in the analysis,
checks and deletes any species for which less than "minim_data" were found}
}
\value{
The result is an object of class dataframe that contains species names
and coordiantes
}
\description{
allows getting occurrences from OBIS and GBIF for a list of species
}
\details{
The function creates downloads occurrences data from the specific sources
and stores them in a data frame.
}
\author{
I. Montero-Serra,  E. Aspillaga, V. Barve, N Barve & K. Kaplan,
}
