% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_rOceans_2_global patterns
%   biodiversity v.0.0.6.r
\name{oceanMaps}
\alias{oceanMaps}
\title{Function #1.4 oceanMaps function in progress to plot different outputs of the rOcean package
visualize spatial patterns in marine biodiversity}
\usage{
oceanMaps(grid, grid_provided = T, logScale = F, occurrences,
  map_occurrences = F, convex_hull = F, long_name = "decimalLongitude",
  lat_name = "decimalLatitude", min_long = -180, max_long = 180,
  min_lat = -90, max_lat = 90, species_name = "species",
  background_color = "grey10", dot_color = "steelblue",
  low_color = "steelblue", mid_color = "gold", high_color = "firebrick",
  col_steps = 20, main = "")
}
\arguments{
\item{grid}{raster of species presences, richness, abundance or other metrics.}

\item{grid_provided}{to indicate whether a raster is provided to be mapped}

\item{occurrences}{species occurrences dataframe with at least latitude and longitude columns}

\item{map_occurrences}{if true, occurrences are mapped for each species}

\item{convex_hull}{if true, a convex hull is computed and mapped for each species of occurrences}

\item{long_name}{column name for longitudes in the "occurrences" dataframe}

\item{lat_name}{column name for latitudes in the "occurrences" dataframe}

\item{min_long}{minimum longitude of the analysis}

\item{max_long}{maximum longitude of the analysis}

\item{max_lat}{maximum latitude of the analysis}

\item{log_scale}{if True grid values are plotted in a log scale, useful for abundance representations}

\item{species_names}{column name for species names in the "occurrences" dataframe}

\item{max_long}{minimum latitute of the analysis}
}
\value{
The result is a map with the spatial trends
}
\description{
to map different metrics computed using functions of the rOcean package
}
\author{
I. Montero-Serra,  E. Aspillaga, V. Barve, N Barve & K. Kaplan,
}
