% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Functions_rOcean_1_biodiversity_for_few_species_v0.1.0.r
\name{oceanRichness}
\alias{oceanRichness}
\title{Function #1.3 oceanRichness creates a richness layer from occurrence data}
\usage{
oceanRichness(occurrences, species_name = "species",
  long_name = "decimalLongitude", lat_name = "decimalLatitude",
  extent = "global", min_long = -180, max_long = 180, min_lat = -90,
  max_lat = 90, cell_size = 5, richness_map = F)
}
\arguments{
\item{occurrences}{species occurrences dataframe, with at least longitude and latitude columns}

\item{long_name}{column name for longitudes in the "occurrences" dataframe}

\item{lat_name}{column name for latitudes in the "occurrences" dataframe}

\item{min_long}{minimum longitude of the analysis}

\item{max_long}{maximum longitude of the analysis}

\item{max_lat}{maximum latitude of the analysis}

\item{cell_size}{interger indicating grid cell size}

\item{richness_map}{plot the resulting richness map}

\item{species_names}{column name for species names in the "occurrences" dataframe}

\item{max_long}{minimum latitute of the analysis}
}
\value{
The result is an object of class raster with richness grid cell
}
\description{
creates a richness raster from multispecies occurrences data
}
\author{
I. Montero-Serra,  E. Aspillaga, V. Barve, N Barve & K. Kaplan,
}
