% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Functions_rOcean_1_biodiversity_for_few_species_v0.1.0.r
\name{presencesRaster}
\alias{presencesRaster}
\title{Function 1.1 oceanDist: Creates a presence / absence raster from occurrence data}
\usage{
presencesRaster(occurrences, long_name = "decimalLongitude",
  lat_name = "decimalLatitude", extent = "global", min_long = -180,
  max_long = 180, min_lat = -90, max_lat = 90, raster_name = "",
  cell_size = 5)
}
\arguments{
\item{occurrences}{species occurrences dataframe, with at least longitude and latitude columns}

\item{long_name}{column name for longitudes in the "occurrences" dataframe}

\item{lat_name}{column name for latitudes in the "occurrences" dataframe}

\item{min_long}{minimum longitude of the analysis}

\item{max_long}{maximum longitude of the analysis}

\item{max_lat}{maximum latitude of the analysis}

\item{raster_name}{to include a name to each created raster}

\item{cell_size}{interger indicating grid cell size}

\item{max_long}{minimum latitute of the analysis}
}
\value{
The result is an object of class raster with presences (1) and absences (0)
}
\description{
creates a presence absence raster from occurrences data
}
\details{
The function creates a presences raster from  occurrences data
}
\author{
I. Montero-Serra,  E. Aspillaga, V. Barve, N Barve & K. Kaplan,
}
